requirements.txt
transformers
torch
pillow
opencv-python
pandas
numpy
scikit-learn
evaluate
tensorboard
torchvision
matplotlib